* changed TSP distances from Integer to Double + cleanUp method for algorithms * fixed bug in GE * minor changes
git-svn-id: svn://svn.code.sf.net/p/jcop/code/trunk@33 ca20af45-4a20-473b-94bf-0ef87cbc3059

+++ jcop/src/cz/cvut/felk/cig/jcop/problem/tsp/inverting/TSPInverting.java
@@ -24,7 +24,7 @@ public class TSPInverting extends TSP {
         super(configFile);
     }
 
-    public TSPInverting(Integer[][] distances) throws ProblemFormatException {
+    public TSPInverting(Double[][] distances) throws ProblemFormatException {
         super(distances);
     }
